Updated the grenade logic
