+++ Assets/Scripts/Weapon/Projectile/GrenadeWeapon.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5bdb53cf8d71a3f43a03f552ae686fa5
+guid: 9aecc659973eccb4f9ac7d51b7651114
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
