+++ UserSettings/EditorUserSettings.asset
@@ -24,10 +24,10 @@ EditorUserSettings:
       value: 055401055c070858555e0e2643775a4413164b7a7d79703475701e36e0e1616a
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 570101545757510358580f7b40770a12104e1b2c782d25327d711c61b6b26d3c
+      value: 54095056570659020b0a5c7741775c44171548292f2c24657b7a4463e1b13568
       flags: 0
     RecentlyUsedSceneGuid-7:
-      value: 54095056570659020b0a5c7741775c44171548292f2c24657b7a4463e1b13568
+      value: 570101545757510358580f7b40770a12104e1b2c782d25327d711c61b6b26d3c
       flags: 0
     RecentlyUsedSceneGuid-8:
       value: 5b520d0503545b0d0c0c0a2715770748154f4d2c7d7d7e627a7d4a35b4e1646a
